traceabort, fix hbonds

+++ BuiltAssets/nodes/analysis/CNA.cpp
@@ -1,17 +1,16 @@
 /// Common Neighbor Analysis
-///  Result mapping;
+///  Result mapping
 ///   0 = Other
 ///   1 = FCC
 ///   2 = HCP
 ///   3 = BCC
 ///   4 = ICO
-
 #include <iostream>
 
 //in enum Conventional Adaptive
 int type = 0;
 //in range 0.1 0.5
-float radius = 0;
+double radius = 0;
 
 //entry
 void Do () {
