+++ mdvis/src/xdrfile/xdrfile_trr.h
@@ -50,11 +50,11 @@ extern "C" {
   /* Read one frame of an open xtc file. If either of x,v,f,box are
      NULL the arrays will be read from the file but not used.  */
   extern int read_trr(XDRFILE *xd,int *natoms,int *step,float *t,float *lambda,
-		      matrix box,float *x,rvec *v,rvec *f);
+		      matrix box,double *x,rvec *v,rvec *f);
 
   /* Write a frame to xtc file */
   extern int write_trr(XDRFILE *xd,int natoms,int step,float t,float lambda,
-		       matrix box,float *x,rvec *v,rvec *f);
+		       matrix box,double *x,rvec *v,rvec *f);
 
   
 #ifdef __cplusplus
